+++ projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-can-activate-child-guard.router.spec.ts
@@ -2,10 +2,15 @@ import { NgxFeatureToggleCanActivateChildGuard } from './ngx-feature-toggle-can-
 import { Route } from '@angular/router';
 import { set } from 'feature-toggle-service';
 
+const fakeRouter = {
+  navigate: () => {},
+} as any;
+
 describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   beforeEach(() => {
     set({ isFirstFeatureEnabled: true, isSecondFeatureEnabled: false });
     spyOn(console, 'error');
+    spyOn(fakeRouter, 'navigate');
   });
 
   afterEach(() => {
@@ -13,7 +18,7 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   });
 
   it('should return `false` if feature toggle is not configured in application level', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
     expect(
       instance.canActivateChild({
         path: 'home',
@@ -24,8 +29,8 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
     ).toBeFalsy();
   });
 
-  it('should return `false` if feature toggle is not added in route', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+  it('should return `false` if feature toggle key does not exist in route', () => {
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     const result = instance.canActivateChild({
       path: 'home',
@@ -33,12 +38,12 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
 
     expect(result).toBeFalsy();
     expect(console.error).toHaveBeenCalledWith(
-      '`NgxFeatureToggleCanActivateChildGuard` need to receive `featureToggle` as data in your route configuration.'
+      '`NgxFeatureToggleCanActivateChildGuard` need to receive `featureToggle` as data as an array or string in your route configuration.'
     );
   });
 
-  it('should return `false` if feature toggle is not added in route', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+  it('should return `false` if feature toggle is not added in route as an array', () => {
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     const result = instance.canActivateChild({
       data: {
@@ -53,7 +58,7 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   });
 
   it('should return `false` if feature toggle is disabled', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     expect(
       instance.canActivateChild({
@@ -64,8 +69,22 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
     ).toBeFalsy();
   });
 
+  it('should return `false` and redirect to the specific URL if feature toggle is disabled AND route contains `redirectTo`', () => {
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
+
+    expect(
+      instance.canActivateChild({
+        data: {
+          featureToggle: ['isSecondFeatureEnabled'],
+          redirectTo: '/redirect-url',
+        },
+      } as Route)
+    ).toBeFalsy();
+    expect(fakeRouter.navigate).toHaveBeenCalledWith(['/redirect-url']);
+  });
+
   it('should NOT console errors if code is running in production mode', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
     spyOn(instance, 'isDevMode').and.returnValue(false);
 
     instance.canActivateChild({
@@ -81,7 +100,7 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   });
 
   it('should return `true` if feature toggle is enabled', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     expect(
       instance.canActivateChild({
@@ -93,7 +112,7 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   });
 
   it('should return `true` if feature toggle is disabled AND route configuration starts with `!`', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     expect(
       instance.canActivateChild({
@@ -105,7 +124,7 @@ describe('Component: NgxFeatureToggleCanActivateChildGuard', () => {
   });
 
   it('should return `true` if combination of feature toggles are thruthy', () => {
-    const instance = new NgxFeatureToggleCanActivateChildGuard();
+    const instance = new NgxFeatureToggleCanActivateChildGuard(fakeRouter);
 
     expect(
       instance.canActivateChild({