+++ projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-can-activate-guard.router.spec.ts
@@ -2,10 +2,15 @@ import { NgxFeatureToggleCanActivateGuard } from './ngx-feature-toggle-can-activ
 import { Route } from '@angular/router';
 import { set } from 'feature-toggle-service';
 
+const fakeRouter = {
+  navigate: () => {},
+} as any;
+
 describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   beforeEach(() => {
     set({ isFirstFeatureEnabled: true, isSecondFeatureEnabled: false });
     spyOn(console, 'error');
+    spyOn(fakeRouter, 'navigate');
   });
 
   afterEach(() => {
@@ -13,7 +18,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should return `false` if feature toggle is not configured in application level', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
     expect(
       instance.canActivate({
         path: 'home',
@@ -32,8 +37,22 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
     ).toBeFalsy();
   });
 
-  it('should return `false` if feature toggle is not added in route', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+  it('should return `false` and redirect to the specific URL if feature toggle is disabled AND route contains `redirectTo`', () => {
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
+
+    expect(
+      instance.canActivate({
+        data: {
+          featureToggle: ['thisFeatureToggleDoesNotExist'],
+          redirectTo: '/redirect-url',
+        },
+      } as Route)
+    ).toBeFalsy();
+    expect(fakeRouter.navigate).toHaveBeenCalledWith(['/redirect-url']);
+  });
+
+  it('should return `false` if feature toggle key does not exist in route', () => {
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     const result = instance.canActivate({
       path: 'home',
@@ -41,12 +60,12 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
 
     expect(result).toBeFalsy();
     expect(console.error).toHaveBeenCalledWith(
-      '`NgxFeatureToggleCanActivateGuard` need to receive `featureToggle` as data in your route configuration.'
+      '`NgxFeatureToggleCanActivateGuard` need to receive `featureToggle` as data as an array or string in your route configuration.'
     );
   });
 
-  it('should return `false` if feature toggle is not added in route', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+  it('should return `false` if feature toggle is not added in route as an array or string', () => {
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     const result = instance.canActivate({
       data: {
@@ -61,7 +80,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should return `false` if feature toggle is disabled', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     expect(
       instance.canActivate({
@@ -81,7 +100,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should NOT console errors if code is running in production mode', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
     spyOn(instance, 'isDevMode').and.returnValue(false);
 
     instance.canActivate({
@@ -97,7 +116,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should return `true` if feature toggle is enabled', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     expect(
       instance.canActivate({
@@ -117,7 +136,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should return `true` if feature toggle is disabled AND route configuration starts with `!`', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     expect(
       instance.canActivate({
@@ -137,7 +156,7 @@ describe('Component: NgxFeatureToggleCanActivateGuard', () => {
   });
 
   it('should return `true` if combination of feature toggles are thruthy', () => {
-    const instance = new NgxFeatureToggleCanActivateGuard();
+    const instance = new NgxFeatureToggleCanActivateGuard(fakeRouter);
 
     expect(
       instance.canActivate({