+++ src/app/restrict-page-due-feature-toggle/restrict-page-due-feature-toggle.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-restrict-page-due-feature-toggle',
+  templateUrl: './restrict-page-due-feature-toggle.component.html',
+  styleUrls: ['./restrict-page-due-feature-toggle.component.scss'],
+})
+export class RestrictPageDueFeatureToggleComponent {
+  constructor() {}
+}