+++ src/app/restrict-page-due-feature-toggle/restrict-page-due-feature-toggle.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RestrictPageDueFeatureToggleComponent } from './restrict-page-due-feature-toggle.component';
+
+@NgModule({
+  imports: [CommonModule],
+  declarations: [RestrictPageDueFeatureToggleComponent],
+})
+export class RestrictPageDueFeatureToggleModule {}